tip clipping new check was introduced

+++ assembler/src/include/levenshtein.hpp
@@ -5,12 +5,12 @@
 /*
  * Little modified copy-paste from http://www.merriampark.com/ldcpp.htm
  */
-inline int edit_distance(const std::string& source, const std::string& target) {
+inline size_t edit_distance(const std::string& source, const std::string& target) {
 
   // Step 1
 
-  const int n = source.length();
-  const int m = target.length();
+  const size_t n = source.length();
+  const size_t m = target.length();
   if (n == 0) {
     return m;
   }
@@ -20,42 +20,42 @@ inline int edit_distance(const std::string& source, const std::string& target) {
 
   // Good form to declare a TYPEDEF
 
-  typedef std::vector< std::vector<int> > Tmatrix;
+  typedef std::vector< std::vector<size_t> > Tmatrix;
 
   Tmatrix matrix(n+1);
 
   // Size the vectors in the 2.nd dimension. Unfortunately C++ doesn't
   // allow for allocation on declaration of 2.nd dimension of vec of vec
 
-  for (int i = 0; i <= n; i++) {
+  for (size_t i = 0; i <= n; i++) {
     matrix[i].resize(m+1);
   }
 
   // Step 2
 
-  for (int i = 0; i <= n; i++) {
+  for (size_t i = 0; i <= n; i++) {
     matrix[i][0]=i;
   }
 
-  for (int j = 0; j <= m; j++) {
+  for (size_t j = 0; j <= m; j++) {
     matrix[0][j]=j;
   }
 
   // Step 3
 
-  for (int i = 1; i <= n; i++) {
+  for (size_t i = 1; i <= n; i++) {
 
     const char s_i = source[i-1];
 
     // Step 4
 
-    for (int j = 1; j <= m; j++) {
+    for (size_t j = 1; j <= m; j++) {
 
       const char t_j = target[j-1];
 
       // Step 5
 
-      int cost;
+      size_t cost;
       if (s_i == t_j) {
         cost = 0;
       }
@@ -65,10 +65,10 @@ inline int edit_distance(const std::string& source, const std::string& target) {
 
       // Step 6
 
-      const int above = matrix[i-1][j];
-      const int left = matrix[i][j-1];
-      const int diag = matrix[i-1][j-1];
-      int cell = min( above + 1, min(left + 1, diag + cost));
+      const size_t above = matrix[i-1][j];
+      const size_t left = matrix[i][j-1];
+      const size_t diag = matrix[i-1][j-1];
+      size_t cell = min( above + 1, min(left + 1, diag + cost));
 
       // Step 6A: Cover transposition, in addition to deletion,
       // insertion and substitution. This step is taken from:
@@ -77,7 +77,7 @@ inline int edit_distance(const std::string& source, const std::string& target) {
       // (http://www.acm.org/~hlb/publications/asm/asm.html)
 
       if (i>2 && j>2) {
-        int trans=matrix[i-2][j-2]+1;
+        size_t trans=matrix[i-2][j-2]+1;
         if (source[i-2]!=t_j) trans++;
         if (s_i!=target[j-2]) trans++;
         if (cell>trans) cell=trans;
